# [jdbc.read] Properly pre-compute column fetching thunk