Thanks to your comment on #179 I finally understood what is the deal with column thinks and such. This PR moves computing the thunk to where it should be – outside of the returned lambda, so that the machinery is not re-run for every row. The benchmark from #185 gives this:

```
master (results are reused):
Time per call: 18.89 ms   Alloc per call: 17,484,278b
Time per call: 18.90 ms   Alloc per call: 17,483,998b
Time per call: 18.87 ms   Alloc per call: 17,483,817b

After PR:
Time per call: 9.83 ms   Alloc per call: 11,968,423b   Iterations: 209
Time per call: 9.99 ms   Alloc per call: 11,968,262b   Iterations: 201
Time per call: 9.76 ms   Alloc per call: 11,968,255b   Iterations: 209
```